Bug Fix: NVARCHAR Ambiguity
NVARCHAR(x) has now been replaced by CHAR(x) CHARACTER SET UTF8MB4, as it will soon replaced CS UTF8MB3. This might be less efficient storage-wise, but is a better choice if we wish to be more inclusive.

Also, updated the ERD to reflect the change and added the ERD to the README file.